backend/requirements.txt
fastapi==0.96.0
uvicorn[standard]==0.23.1
boto3==1.40.33
openai==0.28.0
python-dotenv==1.0.0
numpy>=2.1.3
scikit-learn>=1.5.0
apscheduler==3.10.1
requests>=2.32.4
